Newww/generate_poses.py
"""
Generate pose keypoints for DeepFashion dataset using OpenPose
This script processes all images and extracts pose keypoints
"""

import os
import json
import cv2
import numpy as np
from tqdm import tqdm
import argparse

# Try to import OpenPose (you'll need to install it)
OPENPOSE_AVAILABLE = False
try:
    import pyopenpose as op  # type: ignore
    OPENPOSE_AVAILABLE = True
    print("✅ OpenPose module found!")
except ImportError as e:
    print(f"⚠️  OpenPose not available: {e}")
    print("   Will create synthetic poses instead.")
    print("   To install OpenPose, follow instructions at: https://github.com/CMU-Perceptual-Computing-Lab/openpose")
except Exception as e:
    print(f"⚠️  Error importing OpenPose: {e}")
    print("   Will create synthetic poses instead.")

class PoseGenerator:
    """Generate pose keypoints from images"""
    
    def __init__(self, openpose_path=None):
        self.openpose_available = OPENPOSE_AVAILABLE
        
        if self.openpose_available and openpose_path:
            self.setup_openpose(openpose_path)
        else:
            print("Using synthetic pose generation (OpenPose not available)")
    
    def setup_openpose(self, openpose_path):
        """Setup OpenPose"""
        try:
            # OpenPose parameters
            params = dict()
            params["model_folder"] = openpose_path
            params["net_resolution"] = "256x256"
            params["output_resolution"] = "256x256"
            params["num_gpu"] = 1
            params["num_gpu_start"] = 0
            params["disable_blending"] = False
            params["default_model_folder"] = openpose_path
            
            # Starting OpenPose
            self.opWrapper = op.WrapperPython()
            self.opWrapper.configure(params)
            self.opWrapper.start()
            
            print("✅ OpenPose initialized successfully!")
            
        except Exception as e:
            print(f"❌ Error initializing OpenPose: {e}")
            self.openpose_available = False
    
    def detect_pose_openpose(self, image_path):
        """Detect pose using OpenPose"""
        try:
            # Read image
            datum = op.Datum()
            imageToProcess = cv2.imread(image_path)
            datum.cvInputData = imageToProcess
            
            # Process image
            self.opWrapper.emplaceAndPop([datum])
            
            # Extract keypoints
            if len(datum.poseKeypoints) > 0:
                keypoints = datum.poseKeypoints[0]  # First person
                # Convert to our format (x, y, confidence)
                pose_data = {
                    "version": "1.3",
                    "people": [
                        {
                            "pose_keypoints_2d": keypoints.flatten().tolist(),
                            "face_keypoints_2d": [],
                            "hand_left_keypoints_2d": [],
                            "hand_right_keypoints_2d": [],
                            "pose_keypoints_3d": [],
                            "face_keypoints_3d": [],
                            "hand_left_keypoints_3d": [],
                            "hand_right_keypoints_3d": []
                        }
                    ]
                }
                return pose_data
            else:
                return self.generate_synthetic_pose()
                
        except Exception as e:
            print(f"Error processing {image_path}: {e}")
            return self.generate_synthetic_pose()
    
    def generate_synthetic_pose(self):
        """Generate synthetic pose when OpenPose is not available"""
        # COCO format: 18 keypoints
        # Generate a realistic standing pose with some variation
        base_pose = np.array([
            [0.5, 0.1],   # nose
            [0.45, 0.08], [0.55, 0.08],  # eyes
            [0.42, 0.12], [0.58, 0.12],  # ears
            [0.4, 0.25], [0.6, 0.25],    # shoulders
            [0.4, 0.4], [0.6, 0.4],      # elbows
            [0.4, 0.55], [0.6, 0.55],    # wrists
            [0.45, 0.6], [0.55, 0.6],    # hips
            [0.45, 0.8], [0.55, 0.8],    # knees
            [0.45, 0.95], [0.55, 0.95]   # ankles
        ])
        
        # Add random variation
        noise = np.random.normal(0, 0.02, base_pose.shape)
        pose = base_pose + noise
        pose = np.clip(pose, 0, 1)
        
        # Convert to COCO format (x, y, confidence)
        keypoints = []
        for x, y in pose:
            keypoints.extend([x, y, 1.0])  # confidence = 1.0
        
        pose_data = {
            "version": "1.3",
            "people": [
                {
                    "pose_keypoints_2d": keypoints,
                    "face_keypoints_2d": [],
                    "hand_left_keypoints_2d": [],
                    "hand_right_keypoints_2d": [],
                    "pose_keypoints_3d": [],
                    "face_keypoints_3d": [],
                    "hand_left_keypoints_3d": [],
                    "hand_right_keypoints_3d": []
                }
            ]
        }
        
        return pose_data
    
    def process_image(self, image_path, output_path):
        """Process a single image and save pose data"""
        if self.openpose_available:
            pose_data = self.detect_pose_openpose(image_path)
        else:
            pose_data = self.generate_synthetic_pose()
        
        # Save pose data
        with open(output_path, 'w') as f:
            json.dump(pose_data, f, indent=2)
        
        return pose_data

def get_all_image_files(img_dir):
    """Recursively get all image files from subdirectories"""
    image_files = []
    for root, dirs, files in os.walk(img_dir):
        for file in files:
            if file.lower().endswith(('.jpg', '.jpeg', '.png', '.bmp')):
                # Get relative path from img_dir
                rel_path = os.path.relpath(os.path.join(root, file), img_dir)
                image_files.append(rel_path)
    return image_files

def generate_poses_for_dataset(data_dir, openpose_path=None):
    """Generate pose keypoints for all images in the dataset"""
    
    img_dir = os.path.join(data_dir, 'img')
    pose_dir = os.path.join(data_dir, 'pose')
    
    # Create pose directory
    os.makedirs(pose_dir, exist_ok=True)
    
    # Check if images exist
    if not os.path.exists(img_dir):
        print(f"❌ Image directory not found: {img_dir}")
        return False
    
    # Get all image files recursively
    image_files = get_all_image_files(img_dir)
    
    if not image_files:
        print(f"❌ No images found in {img_dir}")
        return False
    
    print(f"Found {len(image_files)} images to process")
    
    # Initialize pose generator
    pose_gen = PoseGenerator(openpose_path)
    
    # Process all images
    processed = 0
    for image_file in tqdm(image_files, desc="Generating poses"):
        image_path = os.path.join(img_dir, image_file)
        
        # Create output filename with same directory structure
        pose_filename = os.path.splitext(image_file)[0] + '_keypoints.json'
        pose_path = os.path.join(pose_dir, pose_filename)
        
        # Create output directory if it doesn't exist
        os.makedirs(os.path.dirname(pose_path), exist_ok=True)
        
        # Skip if already processed
        if os.path.exists(pose_path):
            continue
        
        try:
            # Process image
            pose_gen.process_image(image_path, pose_path)
            processed += 1
            
        except Exception as e:
            print(f"Error processing {image_file}: {e}")
            continue
    
    print(f"✅ Processed {processed} images")
    print(f"Pose files saved in: {pose_dir}")
    
    return True

def main():
    parser = argparse.ArgumentParser(description='Generate pose keypoints for DeepFashion dataset')
    parser.add_argument('--data_dir', type=str, default='./deepfashion_data',
                       help='Path to DeepFashion dataset')
    parser.add_argument('--openpose_path', type=str, default=None,
                       help='Path to OpenPose installation (optional)')
    
    args = parser.parse_args()
    
    print("Pose Generation for DeepFashion Dataset")
    print("=" * 50)
    
    if not os.path.exists(args.data_dir):
        print(f"❌ Dataset directory not found: {args.data_dir}")
        print("Please run the dataset setup script first.")
        return
    
    # Generate poses
    success = generate_poses_for_dataset(args.data_dir, args.openpose_path)
    
    if success:
        print("\n✅ Pose generation completed!")
        print("You can now run the training script.")
    else:
        print("\n❌ Pose generation failed!")

if __name__ == '__main__':
    main()
